+++ src/cli/ccb-multi-history.ts
@@ -3,7 +3,7 @@
 import * as fs from 'fs';
 import * as path from 'path';
 import chalk from 'chalk';
-import { getProjectInfo } from '../lib/utils';
+import { getProjectInfo } from '../utils';
 
 async function main() {
   try {